refactor(utils): 도구들을 로또 기계 폴더내에 위치
- calculateManager가 안에 들어감으로써
- 도구들이 실물 Lotto와 같은 boundary에 위치하는 것이 직관적으로 와닿지 않다고 판단
- 결국 로또 기계 안에서 쓰이는 도구들이기 때문에 로또 기계 안에 위치
